Fix error in tender record insert.

+++ src/main/java/com/roozen/register/dao/OrderDao.java
@@ -144,6 +144,7 @@ public class OrderDao {
 
     private void insertTenderRecord(Order order) {
         Assert.notNull(order);
+        if (order.getTenderRecord() == null) return;
 
         jdbcTemplate.update("delete from tender_record where order_id = " + order.getOrderId(), new HashMap<>());
